This commit was manufactured by cvs2svn to create tag 'RELENG_1_0_4_RELEASE'.
git-svn-id: https://vrjuggler.googlecode.com/svn/tags/RELENG_1_0_4_RELEASE/juggler_1.0@7378 08b38cba-cd3b-11de-854e-f91c5b6e4272

+++ Config/vjVarValue.cpp
@@ -249,7 +249,7 @@ vjVarValue::operator float () const {
     case T_FLOAT:
         return floatval;
     case T_INT:
-        return (float)floatval;
+        return (float)intval;
     case T_BOOL:
         return (float)boolval;
     case T_INVALID:
